requirements.txt
flask
numpy
opencv-python
opencv-python-headless
torch
torchvision
resnet
Pillow
RealESRGAN
